// src/models/denuncias.ts
export class Denuncias {
  id: number;
  procedimentos: string;
  data: string;
  hora: string;
  origem: string;
  status: string;
  reclamante: string;
  contato_reclamante: string;
  denunciado: string;
  contato_denunciado: string;
  endereco: string;
  bairro: string;
  ponto_de_referencia: string;
  descricao: string;
}